+++ scripts/game/modifiers/VortexModifier.cs
@@ -17,7 +17,7 @@ public class VortexModifier : Modifier, IObjectRenderModifier<Note>
 
         if (depth > 0)
         {
-            depth /= (float)attempt.Settings.ApproachDistance;
+            depth /= (float)(double)attempt.Settings.ApproachDistance;
 
             float rotation = (float)Math.Pow(depth * 15, 1.5) / 60;
             float sine = (float)Math.Sin(attempt.Progress / attempt.Speed / 1000 * attempt.Settings.ApproachTime);
